pyproject.toml
[project]
name = "booklog-to-obsidian"
version = "0.1.0"
description = "ブクログのCSVデータをObsidianに同期するツール"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "watchdog>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = ["tests*"]

[project.scripts]
booklog-sync = "booklog_sync.main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
